Move delete logic to FormField
