+++ front/app/components/FormBuilder/components/FormBuilderSettings/ContentSettings/index.tsx
@@ -2,8 +2,7 @@ import React from 'react';
 import { useFormContext } from 'react-hook-form';
 
 // components
-import { Box, Text, colors } from '@citizenlab/cl2-component-library';
-import Button from 'components/UI/Button';
+import { Box, Text } from '@citizenlab/cl2-component-library';
 import { SectionField } from 'components/admin/Section';
 import InputMultilocWithLocaleSwitcher from 'components/HookForm/InputMultilocWithLocaleSwitcher';
 import Toggle from 'components/HookForm/Toggle';
@@ -21,40 +20,20 @@ import { IFlatCustomFieldWithIndex } from 'api/custom_fields/types';
 import useLocale from 'hooks/useLocale';
 import { isNilOrError } from 'utils/helperUtils';
 import QuillMultilocWithLocaleSwitcher from 'components/HookForm/QuillMultilocWithLocaleSwitcher';
-import { builtInFieldKeys } from 'components/FormBuilder/utils';
 import { get } from 'lodash-es';
 
 type ContentSettingsProps = {
   field: IFlatCustomFieldWithIndex;
-  onDelete: (fieldIndex: number) => void;
-  onClose: () => void;
-  isDeleteDisabled?: boolean;
   locales: Locale[];
 };
 
-export const ContentSettings = ({
-  field,
-  locales,
-  onClose,
-  isDeleteDisabled,
-  onDelete,
-}: ContentSettingsProps) => {
-  const { watch, trigger, setValue } = useFormContext();
+export const ContentSettings = ({ field, locales }: ContentSettingsProps) => {
+  const { watch } = useFormContext();
   const logic = watch(`customFields.${field.index}.logic`);
   const lockedAttributes = field?.constraints?.locks;
   const platformLocale = useLocale();
   const hasRules = logic && logic.rules && logic.rules.length > 0;
   const isFieldGrouping = ['page', 'section'].includes(field.input_type);
-  const handleDelete = () => {
-    if (builtInFieldKeys.includes(field.key)) {
-      const newField = { ...field, enabled: false };
-      setValue(`customFields.${field.index}`, newField);
-      trigger();
-      onClose();
-    } else {
-      onDelete(field.index);
-    }
-  };
 
   const handleKeyDown = (event: React.KeyboardEvent<Element>) => {
     // We want to prevent the form builder from being closed when enter is pressed
@@ -107,36 +86,6 @@ export const ContentSettings = ({
             </SectionField>
           </>
         )}
-        <Box
-          display="flex"
-          justifyContent="space-between"
-          gap="16px"
-          borderTop={`1px solid ${colors.divider}`}
-          pt="36px"
-        >
-          <Button
-            id="e2e-settings-done-button"
-            buttonStyle="secondary"
-            onClick={onClose}
-            minWidth="160px"
-            width="100%"
-          >
-            <FormattedMessage {...messages.done} />
-          </Button>
-          {!get(lockedAttributes, 'enabled', false) && (
-            <Button
-              px="28px"
-              icon="delete"
-              buttonStyle="primary-outlined"
-              borderColor={colors.error}
-              textColor={colors.error}
-              iconColor={colors.error}
-              onClick={handleDelete}
-              data-cy="e2e-delete-field"
-              disabled={isDeleteDisabled}
-            />
-          )}
-        </Box>
       </Box>
     );
   }
