+++ front/app/components/FormBuilder/components/FormFields/FormField/index.tsx
@@ -1,9 +1,16 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { useFormContext, useFieldArray } from 'react-hook-form';
 
 // components
-import { Box, colors } from '@citizenlab/cl2-component-library';
+import {
+  Box,
+  Title,
+  Text,
+  colors,
+  Button,
+} from '@citizenlab/cl2-component-library';
 import MoreActionsMenu from 'components/UI/MoreActionsMenu';
+import Modal from 'components/UI/Modal';
 import FieldTitle from './FieldTitle';
 import Logic from './Logic';
 import IconsAndBadges from './IconsAndBadges';
@@ -15,6 +22,7 @@ import { rgba } from 'polished';
 
 // utils
 import { getFieldBackgroundColor } from '../utils';
+import { builtInFieldKeys } from 'components/FormBuilder/utils';
 
 // Translation
 import { useIntl } from 'utils/cl-intl';
@@ -42,8 +50,7 @@ type Props = {
   selectedFieldId?: string;
   builderConfig: FormBuilderConfig;
   fieldNumbers: Record<string, number>;
-  onDelete: (fieldIndex: number) => void;
-  isDeleteDisabled: boolean;
+  closeSettings: () => void;
 };
 
 export const FormField = ({
@@ -52,18 +59,19 @@ export const FormField = ({
   selectedFieldId,
   builderConfig,
   fieldNumbers,
-  onDelete,
-  isDeleteDisabled,
+  closeSettings,
 }: Props) => {
   const {
     watch,
     formState: { errors },
     trigger,
+    setValue,
   } = useFormContext();
   const { formatMessage } = useIntl();
+  const [showDeleteModal, setShowDeleteModal] = useState(false);
   const formCustomFields: IFlatCustomField[] = watch('customFields');
   const index = formCustomFields.findIndex((f) => f.id === field.id);
-  const { insert } = useFieldArray({
+  const { insert, move, remove } = useFieldArray({
     name: 'customFields',
   });
   const { formEndPageLogicOption, displayBuiltInFields, groupingType } =
@@ -75,6 +83,14 @@ export const FormField = ({
     field.input_type !== 'page' ? field.logic.rules : field.logic;
 
   const isFieldGrouping = ['page', 'section'].includes(field.input_type);
+
+  // Group is only deletable when we have more than one group
+  const isGroupDeletable =
+    formCustomFields.filter((field) => field.input_type === groupingType)
+      .length > 1;
+  const isDeleteDisabled = !(
+    field?.input_type !== groupingType || isGroupDeletable
+  );
 
   const editFieldAndValidate = () => {
     onEditField({ ...field, index });
@@ -131,6 +147,83 @@ export const FormField = ({
     return duplicatedField;
   }
 
+  const onDelete = (fieldIndex: number) => {
+    if (builtInFieldKeys.includes(field.key)) {
+      const newField = { ...field, enabled: false };
+      setValue(`customFields.${index}`, newField);
+    } else {
+      const field = formCustomFields[fieldIndex];
+
+      // When the first group is deleted, it's questions go to the next group
+      if (fieldIndex === 0 && field.input_type === groupingType) {
+        const nextGroupIndex = formCustomFields.findIndex(
+          (field, fieldIndex) =>
+            field.input_type === groupingType && fieldIndex !== 0
+        );
+        move(nextGroupIndex, 0);
+        remove(1);
+      } else {
+        remove(fieldIndex);
+      }
+    }
+    closeSettings();
+    trigger();
+  };
+
+  const closeModal = () => {
+    setShowDeleteModal(false);
+  };
+  const openModal = () => {
+    setShowDeleteModal(true);
+  };
+
+  const deleteField = (fieldIndex: number) => {
+    // Check if deleted field has linked logic
+    const doesPageHaveLinkedLogic = formCustomFields.some((formField) => {
+      if (formField.logic && formField.logic.rules) {
+        return formField.logic.rules.some(
+          (rule) =>
+            rule.goto_page_id === field.id ||
+            rule.goto_page_id === field.temp_id
+        );
+      } else if (formField.logic && formField.logic?.next_page_id) {
+        return (
+          formField.logic?.next_page_id === field.id ||
+          formField.logic?.next_page_id === field.temp_id
+        );
+      }
+      return false;
+    });
+
+    if (doesPageHaveLinkedLogic) {
+      openModal();
+    } else {
+      onDelete(fieldIndex);
+    }
+  };
+
+  const removeLogicAndDelete = () => {
+    formCustomFields.map((formField, i) => {
+      if (formField.logic && formField.logic.rules) {
+        const updatedRules = formField.logic.rules.filter(
+          (rule) =>
+            rule.goto_page_id !== field.id &&
+            rule.goto_page_id !== field.temp_id
+        );
+        setValue(`customFields.${i}.logic.rules`, updatedRules);
+      } else if (formField.logic && formField.logic.next_page_id) {
+        if (
+          formField.logic.next_page_id === field.id ||
+          formField.logic.next_page_id === field.temp_id
+        ) {
+          setValue(`customFields.${i}.logic`, {});
+        }
+      }
+    });
+
+    onDelete(index);
+  };
+
   const actions = [
     ...(field.input_type !== groupingType
       ? [
@@ -151,7 +244,7 @@ export const FormField = ({
           {
             handler: (event: React.MouseEvent) => {
               event.stopPropagation();
-              onDelete(index);
+              deleteField(index);
             },
             label: formatMessage(messages.delete),
             icon: 'delete' as const,
@@ -161,71 +254,105 @@ export const FormField = ({
   ];
 
   return (
-    <FormFieldsContainer
-      role={'button'}
-      key={field.id}
-      background={getFieldBackgroundColor(selectedFieldId, field, hasErrors)}
-      onClick={() => {
-        editFieldAndValidate();
-      }}
-      data-cy="e2e-field-row"
-    >
-      <FlexibleRow rowHeight={isFieldGrouping ? '50px' : '70px'}>
-        <Box
-          display="flex"
-          justifyContent="space-between"
-          className="expand"
-          width="100%"
-          h="100%"
-          alignItems="center"
-          flexWrap="wrap"
-          ml={isFieldGrouping ? '8px' : '32px'}
-        >
-          <Box display="flex" alignItems="center" height="100%" flex="2">
-            <Box display="block">
-              <FieldTitle
-                hasErrors={hasErrors}
-                field={field}
-                fieldNumber={fieldNumbers[field.id]}
-              />
-              {showLogicOnRow && (
-                <Logic
+    <>
+      <FormFieldsContainer
+        role={'button'}
+        key={field.id}
+        background={getFieldBackgroundColor(selectedFieldId, field, hasErrors)}
+        onClick={() => {
+          editFieldAndValidate();
+        }}
+        data-cy="e2e-field-row"
+      >
+        <FlexibleRow rowHeight={isFieldGrouping ? '50px' : '70px'}>
+          <Box
+            display="flex"
+            justifyContent="space-between"
+            className="expand"
+            width="100%"
+            h="100%"
+            alignItems="center"
+            flexWrap="wrap"
+            ml={isFieldGrouping ? '8px' : '32px'}
+          >
+            <Box display="flex" alignItems="center" height="100%" flex="2">
+              <Box display="block">
+                <FieldTitle
+                  hasErrors={hasErrors}
                   field={field}
-                  formCustomFields={formCustomFields}
-                  fieldNumbers={fieldNumbers}
-                  formEndPageLogicOption={formEndPageLogicOption}
+                  fieldNumber={fieldNumbers[field.id]}
                 />
-              )}
+                {showLogicOnRow && (
+                  <Logic
+                    field={field}
+                    formCustomFields={formCustomFields}
+                    fieldNumbers={fieldNumbers}
+                    formEndPageLogicOption={formEndPageLogicOption}
+                  />
+                )}
+              </Box>
+            </Box>
+            <Box
+              flex="1"
+              display="flex"
+              justifyContent="flex-end"
+              alignItems="center"
+            >
+              <IconsAndBadges
+                field={field}
+                displayBuiltInFields={displayBuiltInFields}
+              />
             </Box>
           </Box>
           <Box
-            flex="1"
+            mr="32px"
+            ml="12px"
             display="flex"
-            justifyContent="flex-end"
             alignItems="center"
+            justifyContent="center"
+            h="100%"
           >
-            <IconsAndBadges
-              field={field}
-              displayBuiltInFields={displayBuiltInFields}
+            <MoreActionsMenu
+              showLabel={false}
+              color={colors.textSecondary}
+              actions={actions}
+              onClick={(event) => event.stopPropagation()}
             />
           </Box>
+        </FlexibleRow>
+      </FormFieldsContainer>
+      <Modal opened={showDeleteModal} close={closeModal}>
+        <Box display="flex" flexDirection="column" width="100%" p="20px">
+          <Box mb="40px">
+            <Title variant="h3" color="primary">
+              {formatMessage(messages.deleteFieldWithLogicConfirmationQuestion)}
+            </Title>
+            <Text color="primary" fontSize="l">
+              {formatMessage(messages.deleteResultsInfo)}
+            </Text>
+          </Box>
+          <Box
+            display="flex"
+            flexDirection="row"
+            width="100%"
+            alignItems="center"
+          >
+            <Button
+              icon="delete"
+              data-cy="e2e-confirm-delete-page-and-logic"
+              buttonStyle="delete"
+              width="auto"
+              mr="20px"
+              onClick={removeLogicAndDelete}
+            >
+              {formatMessage(messages.confirmDeleteFieldWithLogicButtonText)}
+            </Button>
+            <Button buttonStyle="secondary" width="auto" onClick={closeModal}>
+              {formatMessage(messages.cancelDeleteButtonText)}
+            </Button>
+          </Box>
         </Box>
-        <Box
-          mr="32px"
-          ml="12px"
-          display="flex"
-          alignItems="center"
-          justifyContent="center"
-          h="100%"
-        >
-          <MoreActionsMenu
-            showLabel={false}
-            color={colors.textSecondary}
-            actions={actions}
-            onClick={(event) => event.stopPropagation()}
-          />
-        </Box>
-      </FlexibleRow>
-    </FormFieldsContainer>
+      </Modal>
+    </>
   );
 };
