+++ front/app/components/FormBuilder/components/messages.ts
@@ -145,15 +145,6 @@ export default defineMessages({
     id: 'app.components.formBuilder.editButtonLabel',
     defaultMessage: 'Edit',
   },
-  deleteFieldWithLogicConfirmationQuestion: {
-    id: 'app.components.formBuilder.deleteFieldWithLogicConfirmationQuestion',
-    defaultMessage:
-      'Deleting this page will also delete the logic associated with it. Are you sure you want to delete it?',
-  },
-  confirmDeleteFieldWithLogicButtonText: {
-    id: 'app.components.formBuilder.confirmDeleteFieldWithLogicButtonText',
-    defaultMessage: 'Yes, delete page',
-  },
   invalidLogicBadgeMessage: {
     id: 'app.components.formBuilder.invalidLogicBadgeMessage',
     defaultMessage: 'Invalid logic',
@@ -165,14 +156,6 @@ export default defineMessages({
   optional: {
     id: 'app.components.formBuilder.optional',
     defaultMessage: 'Optional',
-  },
-  deleteResultsInfo: {
-    id: 'app.components.formBuilder.deleteResultsInfo',
-    defaultMessage: 'This cannot be undone',
-  },
-  cancelDeleteButtonText: {
-    id: 'app.components.formBuilder.cancelDeleteButtonText',
-    defaultMessage: 'Cancel',
   },
   // Logic Settings
   formEnd: {
